Modularization of voice server
