+++ llm_control/__main__.py
@@ -31,7 +31,53 @@ if __name__ == "__main__":
             try:
                 # Import the module lazily to avoid issues with missing dependencies
                 try:
-                    from llm_control import voice_control_server
+                    # Use the modularized version
+                    from llm_control.voice.server import run_server
+                    
+                    # Parse command-line arguments
+                    import argparse
+                    
+                    parser = argparse.ArgumentParser(description='Voice control server')
+                    
+                    parser.add_argument('--host', type=str, default='0.0.0.0',
+                                        help='Host to bind to (default: 0.0.0.0)')
+                    parser.add_argument('--port', type=int, default=5000,
+                                        help='Port to bind to (default: 5000)')
+                    parser.add_argument('--debug', action='store_true',
+                                        help='Enable debug mode')
+                    parser.add_argument('--whisper-model', type=str, default='medium',
+                                        choices=['tiny', 'base', 'small', 'medium', 'large'],
+                                        help='Whisper model size (default: medium)')
+                    parser.add_argument('--ollama-model', type=str, default='llama3.1',
+                                        help='Ollama model to use (default: llama3.1)')
+                    parser.add_argument('--ollama-host', type=str, default='http://localhost:11434',
+                                        help='Ollama API host (default: http://localhost:11434)')
+                    parser.add_argument('--disable-translation', action='store_true',
+                                        help='Disable automatic translation of non-English languages')
+                    parser.add_argument('--language', type=str, default='es',
+                                        help='Default language for voice recognition (default: es)')
+                    parser.add_argument('--disable-screenshots', action='store_true',
+                                        help='Disable capturing screenshots after command execution')
+                    parser.add_argument('--enable-failsafe', action='store_true',
+                                        help='Enable PyAutoGUI failsafe (move mouse to upper-left corner to abort)')
+                    parser.add_argument('--screenshot-dir', type=str, default='.',
+                                        help='Directory where screenshots will be saved (default: current directory)')
+                    
+                    args = parser.parse_args()
+                    
+                    # Update environment variables
+                    os.environ["WHISPER_MODEL_SIZE"] = args.whisper_model
+                    os.environ["OLLAMA_MODEL"] = args.ollama_model
+                    os.environ["OLLAMA_HOST"] = args.ollama_host
+                    os.environ["TRANSLATION_ENABLED"] = "false" if args.disable_translation else "true"
+                    os.environ["DEFAULT_LANGUAGE"] = args.language
+                    os.environ["CAPTURE_SCREENSHOTS"] = "false" if args.disable_screenshots else "true"
+                    os.environ["PYAUTOGUI_FAILSAFE"] = "true" if args.enable_failsafe else "false"
+                    os.environ["SCREENSHOT_DIR"] = args.screenshot_dir
+                    
+                    # Run the server
+                    run_server(host=args.host, port=args.port, debug=args.debug)
+                    
                 except ModuleNotFoundError:
                     logger.error("Voice control server module not found")
                     sys.exit(1)
