+++ llm_control/voice/audio.py
@@ -1,0 +1,175 @@
+"""
+Audio processing module for voice control.
+
+This module handles audio transcription and translation.
+"""
+
+import os
+import sys
+import tempfile
+import logging
+import requests
+from typing import Dict, Any, Optional
+
+# Configure logging
+logger = logging.getLogger("voice-control-audio")
+
+# Import from our modules
+from llm_control.voice.utils import clean_llm_response
+
+# Constants
+DEFAULT_LANGUAGE = os.environ.get("DEFAULT_LANGUAGE", "es")
+WHISPER_MODEL_SIZE = os.environ.get("WHISPER_MODEL_SIZE", "medium")
+OLLAMA_MODEL = os.environ.get("OLLAMA_MODEL", "llama3.1")
+OLLAMA_HOST = os.environ.get("OLLAMA_HOST", "http://localhost:11434")
+
+def transcribe_audio(audio_data, model_size=WHISPER_MODEL_SIZE, language=DEFAULT_LANGUAGE) -> Dict[str, Any]:
+    """
+    Transcribe audio data using Whisper.
+    
+    Args:
+        audio_data: Audio data as bytes
+        model_size: Whisper model size
+        language: Language code
+        
+    Returns:
+        Dictionary with transcription results
+    """
+    try:
+        import whisper
+    except ImportError:
+        return {
+            "error": "Whisper is not installed. Install with 'pip install -U openai-whisper'",
+            "text": ""
+        }
+    
+    try:
+        # Save audio data to a temporary file
+        with tempfile.NamedTemporaryFile(suffix=".wav", delete=False) as temp_file:
+            temp_path = temp_file.name
+            temp_file.write(audio_data)
+        
+        # Load the Whisper model
+        logger.info(f"Loading Whisper model: {model_size}")
+        model = whisper.load_model(model_size)
+        
+        # Transcribe the audio
+        logger.info(f"Transcribing audio with language: {language}")
+        result = model.transcribe(temp_path, language=language)
+        
+        # Get the transcription
+        text = result.get("text", "").strip()
+        detected_language = result.get("language", "unknown")
+        
+        logger.info(f"Transcription: {text} (language: {detected_language})")
+        
+        return {
+            "text": text,
+            "language": detected_language
+        }
+    
+    except Exception as e:
+        logger.error(f"Error transcribing audio: {str(e)}")
+        import traceback
+        logger.error(traceback.format_exc())
+        
+        return {
+            "error": str(e),
+            "text": ""
+        }
+    
+    finally:
+        # Clean up temporary file
+        if 'temp_path' in locals() and os.path.exists(temp_path):
+            try:
+                os.unlink(temp_path)
+            except (OSError, PermissionError):
+                pass
+
+def translate_text(text, model=OLLAMA_MODEL, ollama_host=OLLAMA_HOST) -> Optional[str]:
+    """
+    Translate text from any language to English using Ollama.
+    
+    Args:
+        text: Text to translate
+        model: Ollama model to use
+        ollama_host: Ollama API host
+        
+    Returns:
+        Translated text or None if translation failed
+    """
+    logger.info(f"Translating with Ollama ({model}): {text}")
+    
+    try:
+        # Check if Ollama is running
+        try:
+            response = requests.get(f"{ollama_host}/api/tags", timeout=2)
+            if response.status_code != 200:
+                logger.error(f"Ollama server not responding at {ollama_host}")
+                return None
+        except requests.exceptions.RequestException:
+            logger.error(f"Ollama server not available at {ollama_host}")
+            return None
+        
+        # Prepare the prompt for translation
+        prompt = f"""
+        Translate the following text to English.
+        
+        CRITICAL: DO NOT translate any of the following:
+        1. Proper nouns, UI element names, button labels, or technical terms
+        2. Menu items, tabs, or buttons (like "Actividades", "Archivo", "Configuración")
+        3. Application names (like "Firefox", "Chrome", "Terminal")
+        4. Text inside quotes (e.g., "Hola mundo")
+        5. Any word that might be a desktop element or application name
+        
+        EXAMPLES of words to KEEP in original language:
+        - "actividades" should stay as "actividades" (NEVER translate to "activities")
+        - "opciones" should stay as "opciones" (NEVER translate to "options")
+        - "archivo" should stay as "archivo" (NEVER translate to "file")
+        - "nueva pestaña" should stay as "nueva pestaña" (NEVER translate to "new tab")
+        
+        Spanish → English examples with preserved text:
+        - "haz clic en el botón Cancelar" → "click on the Cancelar button"
+        - "escribe 'Hola mundo' en el campo Mensaje" → "type 'Hola mundo' in the Mensaje field"
+        - "presiona enter en la ventana Configuración" → "press enter in the Configuración window"
+        - "selecciona Archivo desde el menú" → "select Archivo from the menu"
+        - "mueve el cursor a actividades" → "move the cursor to actividades"
+        
+        ```
+        {text}
+        ```
+        
+        RETURN ONLY THE TRANSLATED TEXT - NOTHING ELSE. NO EXPLANATIONS. NO HEADERS. NO NOTES.
+        """
+        
+        # Make API request to Ollama
+        response = requests.post(
+            f"{ollama_host}/api/generate",
+            json={
+                "model": model,
+                "prompt": prompt,
+                "stream": False
+            },
+            timeout=30
+        )
+        
+        if response.status_code != 200:
+            logger.error(f"Error from Ollama API: {response.status_code}")
+            return None
+        
+        # Parse response
+        result = response.json()
+        translated_text = result["response"].strip()
+        
+        # Clean up the response
+        translated_text = clean_llm_response(translated_text)
+        
+        logger.info(f"Translated: {text} → {translated_text}")
+        
+        return translated_text
+    
+    except Exception as e:
+        logger.error(f"Error translating with Ollama: {str(e)}")
+        import traceback
+        logger.error(traceback.format_exc())
+        return None
